fix/menu-events
Fixed it so you can choose to view requests or appointments events directly from the menu.
Removed the ability to view all events at once because it would be to much on screen at once.
Also removed view all entries in the ViewEvents method because we thought it would be better to only view
user specific entries and not everybodys entries at once.