add/accept-or-deny-request-create-personnel-bool
Finished accept/deny request, if a request is accepted/ denied it will remove the request from eventsList.
When accepted the admin will get a create account menu and enter the information for the users account.
Also Changed the CreateAccount method to return a bool instead of void so we can confirm/ not confirm account creation
so the program knows if the request event should be removed from eventList or not.